Use cuew.template.c/h files to generate code
These are easier to edit for changes like the next commit, inserting
autogenerated code in a few places rather than generating all C code
through Python.
