+++ auto/cuew_gen.sh
@@ -6,5 +6,5 @@
 SCRIPT=`realpath -s $0`
 DIR=`dirname $SCRIPT`
 
-python ${DIR}/cuew_gen.py hdr $@ > $DIR/../include/cuew.h
-python ${DIR}/cuew_gen.py impl $@ > $DIR/../src/cuew.c
+python3 ${DIR}/cuew_gen.py hdr $@ > $DIR/../include/cuew.h
+python3 ${DIR}/cuew_gen.py impl $@ > $DIR/../src/cuew.c
